refactor(sections): extract the prose-paragraph backfill into backfill.rs
Pure motion, and the last of the split: sections/mod.rs now holds the
driver — ParsedHelp, parse_with_profile and parse_body — and nothing else.
